requirements.txt
requests
django
Pillow
